+++ examples/storybook/stories/TestCase/ImagePluginWithWidth.tsx
@@ -11,8 +11,9 @@ const ImagePluginStory = () => {
   const [viewerKey, setViewerKey] = useState(1);
 
   const experiments = parseExperiments({
-    ['specs.ricos.skipImageThumbnail']: 'true',
-    ['specs.ricos.imageThumbnailQuality']: '20',
+    // ['specs.ricos.skipImageThumbnail']: 'true',
+    // ['specs.ricos.imageThumbnailQuality']: '20',
+    ['specs.ricos.useQualityPreoad']: 'true',
   });
   console.log({ experiments });
   const _rcProps = {
@@ -20,9 +21,9 @@ const ImagePluginStory = () => {
   };
 
   return (
-    <Page title="Image Plugin With Skip Thumbnail experiment">
+    <Page title="Image Plugin With Thumbnail experiment">
       <ActionButton
-        text={`${experimntEnabled ? 'Disable' : 'Enable'} "Skip Thumbnail" Experiment`}
+        text={`${experimntEnabled ? 'Disable' : 'Enable'} "Quality Thumbnail" Experiment`}
         onClick={() => {
           setExperimentEnabled(!experimntEnabled);
           setViewerKey(viewerKey + 1);
@@ -32,7 +33,7 @@ const ImagePluginStory = () => {
       <div style={{ width: 700 }}>
         <RicosViewer
           key={viewerKey}
-          width={700}
+          // width={700}
           content={imageContentState}
           plugins={[pluginImage()]}
           {...(experimntEnabled && { _rcProps })}